fix: show commit version correctly
