+++ src/cli/tact/index.ts
@@ -1,7 +1,5 @@
-import { readFileSync } from "fs";
-import { basename, dirname, join, normalize, resolve } from "path";
-import { execFileSync } from "child_process";
-import { z, ZodError } from "zod";
+import { basename, dirname, normalize, resolve } from "path";
+import { ZodError } from "zod";
 import { createNodeFileSystem } from "../../vfs/createNodeFileSystem";
 import { createVirtualFileSystem } from "../../vfs/createVirtualFileSystem";
 import { parseAndEvalExpression } from "../../optimizer/interpreter";
@@ -18,6 +16,7 @@ import { build } from "../../pipeline/build";
 import { TactErrorCollection } from "../../error/errors";
 import files from "../../stdlib/stdlib";
 import { cwd } from "process";
+import { getVersion, showCommit } from "../version";
 
 export const main = async () => {
     const Log = CliLogger();
@@ -46,7 +45,7 @@ const processArgs = async (Errors: CliErrors, argv: string[]) => {
 
         await parseArgs(Errors, Args);
     } else {
-        showHelp();
+        await showHelp();
     }
 };
 
@@ -64,7 +63,7 @@ const ArgSchema = (Parser: ArgParser) => {
         .add(Parser.immediate).end;
 };
 
-const showHelp = () => {
+const showHelp = async () => {
     console.log(`Usage
 $ tact [...flags] (--config CONFIG | FILE)
 
@@ -81,7 +80,7 @@ Flags
 
 Examples
   $ tact --version
-  ${getVersion()}
+  ${await getVersion()}
 
 Learn more about Tact:        https://docs.tact-lang.org
 Join Telegram group:          https://t.me/tactlang
@@ -92,22 +91,23 @@ type Args = ArgConsumer<GetParserResult<ReturnType<typeof ArgSchema>>>;
 
 const parseArgs = async (Errors: CliErrors, Args: Args) => {
     if (Args.single("help")) {
-        if (noUnknownParams(Errors, Args)) {
-            showHelp();
+        if (await noUnknownParams(Errors, Args)) {
+            await showHelp();
         }
         return;
     }
 
     if (Args.single("version")) {
-        if (noUnknownParams(Errors, Args)) {
-            showVersion();
+        if (await noUnknownParams(Errors, Args)) {
+            console.log(await getVersion());
+            showCommit();
         }
         return;
     }
 
     const expression = Args.single("eval");
     if (expression) {
-        if (noUnknownParams(Errors, Args)) {
+        if (await noUnknownParams(Errors, Args)) {
             evaluate(expression);
         }
         return;
@@ -146,8 +146,8 @@ const parseArgs = async (Errors: CliErrors, Args: Args) => {
         return;
     }
 
-    if (noUnknownParams(Errors, Args)) {
-        showHelp();
+    if (await noUnknownParams(Errors, Args)) {
+        await showHelp();
     }
 };
 
@@ -228,7 +228,7 @@ const compile = async (
 
     const stdlib = createVirtualFileSystem("@stdlib", files);
 
-    if (noUnknownParams(Errors, Args)) {
+    if (await noUnknownParams(Errors, Args)) {
         // TODO: all flags on the cli should take precedence over flags in the config
         // Make a nice model for it instead of the current mess
         // Consider making overwrites right here or something.
@@ -306,7 +306,10 @@ const setConfigOptions = (config: Config, options: ExtraOptions): void => {
     }
 };
 
-const noUnknownParams = (Errors: CliErrors, Args: Args): boolean => {
+const noUnknownParams = async (
+    Errors: CliErrors,
+    Args: Args,
+): Promise<boolean> => {
     const leftoverArgs = Args.leftover();
 
     if (leftoverArgs.length === 0) {
@@ -316,23 +319,8 @@ const noUnknownParams = (Errors: CliErrors, Args: Args): boolean => {
     for (const argument of leftoverArgs) {
         Errors.unexpectedArgument(argument);
     }
-    showHelp();
+    await showHelp();
     return false;
-};
-
-const showVersion = () => {
-    console.log(getVersion());
-    // if working inside a git repository
-    // also print the current git commit hash
-    try {
-        const gitCommit = execFileSync("git", ["rev-parse", "HEAD"], {
-            encoding: "utf8",
-            stdio: ["ignore", "pipe", "ignore"],
-        }).trim();
-        console.log(`git commit: ${gitCommit}`);
-    } finally {
-        process.exit(0);
-    }
 };
 
 const evaluate = (expression: string) => {
@@ -347,18 +335,4 @@ const evaluate = (expression: string) => {
             process.exit(30);
         }
     }
-};
-
-const getVersion = () => {
-    const packageSchema = z.object({
-        version: z.string(),
-    });
-
-    const packageJsonPath = join(__dirname, "..", "..", "..", "package.json");
-
-    const pkg = packageSchema.parse(
-        JSON.parse(readFileSync(packageJsonPath, "utf-8")),
-    );
-
-    return pkg.version;
 };
